feat(dsv4): store A5 indexer KV in bfloat16
Keep the same A5 --kv-cache-dtype bfloat16 switch, but store indexer K
as bfloat16 instead of float16 so it matches attention KV. Page size is
unchanged (2 bytes/element, no scale). Unquant TopK still uses
compressed key lengths and defaultMask.

+++ tests/ut/worker/test_attn_utils_v2.py
@@ -106,7 +106,7 @@ def test_sfa_indexer_cache_spec_uses_dcp_replication(monkeypatch, replicated_ind
         (
             AscendDeviceType.A5,
             "bfloat16",
-            torch.float16,
+            torch.bfloat16,
             None,
             (128,),
         ),
